bluemoji: extracted element into its own tsx file
